# log: allow --vmodule to downgrade the log level