@MariusVanDerWijden 2025-04-01
## Benchmark:
```
func BenchmarkGlog(b *testing.B) {
	out := new(bytes.Buffer)
	glog := NewGlogHandler(NewTerminalHandlerWithLevel(out, LevelTrace, false))
	glog.Verbosity(LevelCrit)
	logger := NewLogger(glog)
	logger.Warn("This should not be seen", "ignored", "true")
	//glog.Vmodule("logger_test.go=5")
	for i := 0; i < b.N; i++ {
		logger.Trace("a message", "foo", "bar")
	}
}
```

## This PR:
With vmodule disabled:
`BenchmarkGlog-14    	370532622	         3.763 ns/op	       0 B/op	       0 allocs/op`
With vmodule enabled:
`BenchmarkGlog-14    	 2653554	       406.2 ns/op	     151 B/op	       0 allocs/op`

## Master:
With vmodule disabled:
`BenchmarkGlog-14    	307312317	         3.775 ns/op	       0 B/op	       0 allocs/op`
With vmodule enabled:
`BenchmarkGlog-14    	 2882396	       436.6 ns/op	     279 B/op	       0 allocs/op`

So it seems there's almost no performance difference between this PR and master