@fjl 2025-04-02
I guess this works because the logger calls `Enabled` first to check for the fast path.