@MariusVanDerWijden 2025-05-29
```
Master:
BenchmarkVmodule-14    	 1413564	       960.7 ns/op	     569 B/op	       0 allocs/op
PR:
BenchmarkVmodule-14    	 1467424	       837.8 ns/op	     548 B/op	       0 allocs/op
```

```go
func BenchmarkVmodule(b *testing.B) {
	out := new(bytes.Buffer)
	glog := NewGlogHandler(NewTerminalHandlerWithLevel(out, LevelTrace, false))
	glog.Verbosity(LevelCrit)
	logger := NewLogger(glog)
	glog.Vmodule("logger_test.go=5")
	for i := 0; i < b.N; i++ {
		logger.Warn("This should not be seen", "ignored", "true")
		logger.Trace("a message", "foo", "bar")
	}
}
```