@MariusVanDerWijden 2026-01-19
closed in favor of https://github.com/ethereum/go-ethereum/pull/33111